oauth/pyproject.toml
[project]
name = "user_oauth"
version = "0.1.0"
description = "Salesforce Account CRUD sample"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "simple-salesforce>=1.12.6",
    "python-dotenv>=1.0.0",
    "pyjwt[cryptography]>=2.12.1",
]
